Removed old processingChainManager
