+++ server/modules/analysis/manager/calculation/index.js
@@ -1,1 +1,94 @@
-export { fetchCalculationAttributeUuids, updateCalculation } from '../../repository/calculation'
+import * as R from 'ramda'
+import * as DB from '../../../../db'
+
+import * as Chain from '../../../../../common/analysis/processingChain'
+import * as Step from '../../../../../common/analysis/processingStep'
+import * as Calculation from '../../../../../common/analysis/processingStepCalculation'
+import * as ChainValidator from '../../../../../common/analysis/processingChainValidator'
+import { TableCalculation, TableChain } from '../../../../../common/model/db'
+import * as ActivityLog from '../../../../../common/activityLog/activityLog'
+
+import { markSurveyDraft } from '../../../survey/repository/surveySchemaRepositoryUtils'
+import * as ActivityLogRepository from '../../../activityLog/repository/activityLogRepository'
+import * as ChainRepository from '../../repository/chain'
+import * as StepRepository from '../../repository/step'
+import * as CalculationRepository from '../../repository/calculation'
+
+// ====== CREATE
+const _insertCalculation = async ({ user, surveyId, chain, calculation }, client) => {
+  const calculationDb = await CalculationRepository.insertCalculation({ surveyId, calculation }, client)
+  const content = {
+    ...calculationDb,
+    [ActivityLog.keysContent.processingChainUuid]: Chain.getUuid(chain),
+  }
+  const type = ActivityLog.type.processingStepCalculationCreate
+  return ActivityLogRepository.insert(user, surveyId, type, content, false, client)
+}
+
+// ====== READ
+export const { fetchCalculationAttributeUuids } = CalculationRepository
+
+// ====== UPDATE
+export const { updateCalculation } = CalculationRepository
+
+const _updateCalculation = async ({ user, surveyId, chain, calculation, calculationDb }, client) => {
+  const nodeDefUuid = Calculation.getNodeDefUuid(calculation)
+  const nodeDefUuidDb = Calculation.getNodeDefUuid(calculationDb)
+  const propsDiff = Calculation.getPropsDiff(calculation)(calculationDb)
+
+  const fields = {}
+  if (nodeDefUuid !== nodeDefUuidDb) fields[TableCalculation.columnSet.nodeDefUuid] = nodeDefUuid
+  if (!R.isEmpty(propsDiff)) fields[TableCalculation.columnSet.props] = propsDiff
+
+  if (!R.isEmpty(fields)) {
+    const params = { surveyId, calculationUuid: Calculation.getUuid(calculation), fields }
+    const calculationUpdated = await CalculationRepository.updateCalculation(params, client)
+    const content = {
+      ...calculationUpdated,
+      [ActivityLog.keysContent.processingChainUuid]: Chain.getUuid(chain),
+    }
+    const type = ActivityLog.type.processingStepCalculationUpdate
+    return ActivityLogRepository.insert(user, surveyId, type, content, false, client)
+  }
+  return null
+}
+
+// ====== PERSIST
+export const persistCalculation = async ({ user, surveyId, chain, calculation }, client) => {
+  const params = { surveyId, calculationUuid: Calculation.getUuid(calculation) }
+  const calculationDb = await CalculationRepository.fetchCalculation(params, client)
+  return calculationDb
+    ? _updateCalculation({ user, surveyId, chain, calculation, calculationDb }, client)
+    : _insertCalculation({ user, surveyId, chain, calculation }, client)
+}
+
+// ====== DELETE
+export const deleteCalculation = async ({ user, surveyId, stepUuid, calculationUuid }, client = DB.client) =>
+  client.tx(async (tx) => {
+    const [step, calculation] = await Promise.all([
+      StepRepository.fetchStep({ surveyId, stepUuid }, tx),
+      CalculationRepository.deleteCalculation({ surveyId, calculationUuid }, tx),
+      markSurveyDraft(surveyId, tx),
+    ])
+    const chainUuid = Step.getProcessingChainUuid(step)
+
+    // insert activity log
+    const content = {
+      [ActivityLog.keysContent.uuid]: stepUuid,
+      [ActivityLog.keysContent.processingChainUuid]: chainUuid,
+      [ActivityLog.keysContent.processingStepUuid]: stepUuid,
+      [ActivityLog.keysContent.processingStepIndex]: Step.getIndex(step),
+      [ActivityLog.keysContent.index]: Calculation.getIndex(calculation),
+      [ActivityLog.keysContent.labels]: Calculation.getLabels(calculation),
+    }
+    const type = ActivityLog.type.processingStepCalculationDelete
+    await ActivityLogRepository.insert(user, surveyId, type, content, false, tx)
+
+    // Update step validation
+    const chain = await ChainRepository.fetchChain({ surveyId, chainUuid, includeStepsAndCalculations: true }, tx)
+    const stepValidation = await ChainValidator.validateStep(Chain.getStepByIdx(Step.getIndex(step))(chain))
+    const chainUpdated = Chain.assocItemValidation(stepUuid, stepValidation)(chain)
+    // Update processing_chain validation
+    const fields = { [TableChain.columnSet.validation]: Chain.getValidation(chainUpdated) }
+    return ChainRepository.updateChain({ surveyId, chainUuid, fields, dateModified: true }, tx)
+  })
