+++ server/modules/analysis/manager/chain/index.js
@@ -1,20 +1,24 @@
 import * as DB from '../../../../db'
 
-import * as ProcessingChain from '../../../../../common/analysis/processingChain'
-import * as ActivityLog from '../../../../../common/activityLog/activityLog'
+import * as Chain from '../../../../../common/analysis/processingChain'
 import { TableChain } from '../../../../../common/model/db'
+import * as ActivityLog from '../../../../../common/activityLog/activityLog'
 
-import * as ChainRepository from '../../repository/chain'
+import { markSurveyDraft } from '../../../survey/repository/surveySchemaRepositoryUtils'
 import * as ActivityLogRepository from '../../../activityLog/repository/activityLogRepository'
+import * as ChainRepository from '../../repository/chain'
 
-export {
-  countChains,
-  fetchChains,
-  fetchChain,
-  updateChain,
-  removeChainCycles,
-  deleteChain,
-} from '../../repository/chain'
+// ====== CREATE
+const _insertChain = async ({ user, surveyId, chain }, client) => {
+  const chainDb = await ChainRepository.insertChain({ surveyId, chain }, client)
+  await ActivityLogRepository.insert(user, surveyId, ActivityLog.type.processingChainCreate, chainDb, false, client)
+}
+
+// ====== READ
+export const { countChains, fetchChains, fetchChain } = ChainRepository
+
+// ====== UPDATE
+export const { updateChain, removeChainCycles } = ChainRepository
 
 export const updateChainStatusExec = async ({ user, surveyId, chainUuid, statusExec }) =>
   DB.client.tx(async (tx) => {
@@ -24,7 +28,7 @@ export const updateChainStatusExec = async ({ user, surveyId, chainUuid, statusE
         tx
       ),
     ]
-    if (statusExec === ProcessingChain.statusExec.success) {
+    if (statusExec === Chain.statusExec.success) {
       const type = ActivityLog.type.processingChainStatusExecSuccess
       const content = { [ActivityLog.keysContent.uuid]: chainUuid }
       promises.push(ActivityLogRepository.insert(user, surveyId, type, content, false, tx))
@@ -32,3 +36,48 @@ export const updateChainStatusExec = async ({ user, surveyId, chainUuid, statusE
 
     return Promise.all(promises)
   })
+
+const _updateChain = async ({ user, surveyId, chain, chainDb }, client) => {
+  const chainUuid = Chain.getUuid(chain)
+  const propsToUpdate = Chain.getPropsDiff(chain)(chainDb)
+  // activity log for each updated prop
+  const promises = Object.entries(propsToUpdate).map(([key, value]) => {
+    const content = { [ActivityLog.keysContent.uuid]: chainUuid, key, value }
+    const type = ActivityLog.type.processingChainPropUpdate
+    return ActivityLogRepository.insert(user, surveyId, type, content, false, client)
+  })
+  // chain props and validation update
+  const fields = {
+    [TableChain.columnSet.props]: propsToUpdate,
+    [TableChain.columnSet.validation]: Chain.getValidation(chain),
+  }
+  const params = { surveyId, chainUuid, dateModified: true, fields }
+  promises.push(ChainRepository.updateChain(params, client))
+
+  return Promise.all(promises)
+}
+
+// ====== PERSIST
+export const persistChain = async ({ user, surveyId, chain }, client) => {
+  const chainDb = await ChainRepository.fetchChain({ surveyId, chainUuid: Chain.getUuid(chain) }, client)
+  return chainDb
+    ? _updateChain({ user, surveyId, chain, chainDb }, client)
+    : _insertChain({ user, surveyId, chain }, client)
+}
+
+// ====== DELETE
+export const deleteChain = async ({ user, surveyId, chainUuid }, client = DB.client) =>
+  client.tx(async (tx) => {
+    const chains = await ChainRepository.deleteChain({ surveyId, chainUuid }, tx)
+    const content = {
+      [ActivityLog.keysContent.uuid]: chainUuid,
+      [ActivityLog.keysContent.labels]: Chain.getLabels(chains[0]),
+    }
+    return Promise.all([
+      ActivityLogRepository.insert(user, surveyId, ActivityLog.type.processingChainDelete, content, false, tx),
+      markSurveyDraft(surveyId, tx),
+    ])
+  })
+
+export const deleteChainWithoutCycle = async ({ surveyId }, client = DB.client) =>
+  ChainRepository.deleteChain({ surveyId, noCycle: true }, client)
